requirements.txt
# Core libraries
numpy
opencv-python
PyYAML

# Optional utilities
matplotlib
pandas
ultralytics
pillow
scikit-learn